hotfix(DeviceBattery): changes from int to float for battery levels

+++ src/Objects/DeviceBattery.php
@@ -19,7 +19,7 @@ class DeviceBattery
 
     
     public function __construct(
-        public int $level,
+        public float $level,
         public string $status,
     ) {
     }
